Network.AWS.Organizations.CancelHandshake

Description

Cancels a handshake. Canceling a handshake sets the handshake state to CANCELED .

This operation can be called only from the account that originated the handshake. The recipient of the handshake can't cancel it, but can use DeclineHandshake instead. After a handshake is canceled, the recipient can no longer respond to that handshake.

After you cancel a handshake, it continues to appear in the results of relevant APIs for only 30 days. After that it is deleted.

Creating a Request

cancelHandshake

Arguments

:: Text
chHandshakeId

-> CancelHandshake

Creates a value of CancelHandshake with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chHandshakeId - The unique identifier (ID) of the handshake that you want to cancel. You can get the ID from the ListHandshakesForOrganization operation. The regex pattern for handshake ID string requires "h-" followed by from 8 to 32 lower-case letters or digits.

data CancelHandshake

See: cancelHandshake smart constructor.

Request Lenses

chHandshakeId :: Lens' CancelHandshake Text

The unique identifier (ID) of the handshake that you want to cancel. You can get the ID from the ListHandshakesForOrganization operation. The regex pattern for handshake ID string requires "h-" followed by from 8 to 32 lower-case letters or digits.

Destructuring the Response

cancelHandshakeResponse

Arguments

:: Int
chrsResponseStatus

-> CancelHandshakeResponse

Creates a value of CancelHandshakeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

chrsHandshake - A structure that contains details about the handshake that you canceled.

chrsResponseStatus - -- | The response status code.

data CancelHandshakeResponse

See: cancelHandshakeResponse smart constructor.

Response Lenses

chrsHandshake :: Lens' CancelHandshakeResponse (Maybe Handshake)

A structure that contains details about the handshake that you canceled.

chrsResponseStatus :: Lens' CancelHandshakeResponse Int

- | The response status code.