Network.AWS.Organizations.CreateAccount

Description

Creates an AWS account that is automatically a member of the organization whose credentials made the request. This is an asynchronous request that AWS performs in the background. If you want to check the status of the request later, you need the OperationId response element from this operation to provide as a parameter to the DescribeCreateAccountStatus operation.

AWS Organizations preconfigures the new member account with a role (named OrganizationAccountAccessRole by default) that grants administrator permissions to the new account. Principals in the master account can assume the role. AWS Organizations clones the company name and address information for the new account from the organization's master account.

For more information about creating accounts, see Creating an AWS Account in Your Organization in the AWS Organizations User Guide .

Important: You cannot remove accounts that are created with this operation from an organization. That also means that you cannot delete an organization that contains an account that is created with this operation.

This operation can be called only from the organization's master account.

Creating a Request

createAccount

Arguments

:: Text
caEmail

-> Text
caAccountName

-> CreateAccount

Creates a value of CreateAccount with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

caIAMUserAccessToBilling - If set to ALLOW , the new account enables IAM users to access account billing information if they have the required permissions. If set to DENY , then only the root user of the new account can access account billing information. For more information, see Activating Access to the Billing and Cost Management Console in the AWS Billing and Cost Management User Guide . If you do not specify this parameter, the value defaults to ALLOW, and IAM users and roles with the required permissions can access billing information for the new account.

caRoleName - (Optional) The name of an IAM role that Organizations automatically preconfigures in the new member account. This role trusts the master account, allowing users in the master account to assume the role, as permitted by the master account administrator. The role has administrator permissions in the new member account. If you do not specify this parameter, the role name defaults to OrganizationAccountAccessRole . For more information about how to use this role to access the member account, see Accessing and Administering the Member Accounts in Your Organization in the AWS Organizations User Guide , and steps 2 and 3 in Tutorial: Delegate Access Across AWS Accounts Using IAM Roles in the IAM User Guide . The regex pattern that is used to validate this parameter is a string of characters that can consist of uppercase letters, lowercase letters, digits with no spaces, and any of the following characters: =,.@-

caEmail - The email address of the owner to assign to the new member account. This email address must not already be associated with another AWS account.

caAccountName - The friendly name of the member account.

data CreateAccount

See: createAccount smart constructor.

Request Lenses

caIAMUserAccessToBilling :: Lens' CreateAccount (Maybe IAMUserAccessToBilling)

If set to ALLOW , the new account enables IAM users to access account billing information if they have the required permissions. If set to DENY , then only the root user of the new account can access account billing information. For more information, see Activating Access to the Billing and Cost Management Console in the AWS Billing and Cost Management User Guide . If you do not specify this parameter, the value defaults to ALLOW, and IAM users and roles with the required permissions can access billing information for the new account.

caRoleName :: Lens' CreateAccount (Maybe Text)

(Optional) The name of an IAM role that Organizations automatically preconfigures in the new member account. This role trusts the master account, allowing users in the master account to assume the role, as permitted by the master account administrator. The role has administrator permissions in the new member account. If you do not specify this parameter, the role name defaults to OrganizationAccountAccessRole . For more information about how to use this role to access the member account, see Accessing and Administering the Member Accounts in Your Organization in the AWS Organizations User Guide , and steps 2 and 3 in Tutorial: Delegate Access Across AWS Accounts Using IAM Roles in the IAM User Guide . The regex pattern that is used to validate this parameter is a string of characters that can consist of uppercase letters, lowercase letters, digits with no spaces, and any of the following characters: =,.@-

caEmail :: Lens' CreateAccount Text

The email address of the owner to assign to the new member account. This email address must not already be associated with another AWS account.

caAccountName :: Lens' CreateAccount Text

The friendly name of the member account.

Destructuring the Response

createAccountResponse

Arguments

:: Int
carsResponseStatus

-> CreateAccountResponse

Creates a value of CreateAccountResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

carsCreateAccountStatus - A structure that contains details about the request to create an account. This response structure might not be fully populated when you first receive it because account creation is an asynchronous process. You can pass the returned CreateAccountStatus ID as a parameter to DescribeCreateAccountStatus to get status about the progress of the request at later times.

carsResponseStatus - -- | The response status code.

data CreateAccountResponse

See: createAccountResponse smart constructor.

Response Lenses

carsCreateAccountStatus :: Lens' CreateAccountResponse (Maybe CreateAccountStatus)

A structure that contains details about the request to create an account. This response structure might not be fully populated when you first receive it because account creation is an asynchronous process. You can pass the returned CreateAccountStatus ID as a parameter to DescribeCreateAccountStatus to get status about the progress of the request at later times.

carsResponseStatus :: Lens' CreateAccountResponse Int

- | The response status code.