Network.AWS.Organizations.CreateOrganization

Description

Creates an AWS organization. The account whose user is calling the CreateOrganization operation automatically becomes the master account of the new organization.

This operation must be called using credentials from the account that is to become the new organization's master account. The principal must also have the relevant IAM permissions.

By default (or if you set the FeatureSet parameter to ALL ), the new organization is created with all features enabled and service control policies automatically enabled in the root. If you instead choose to create the organization supporting only the consolidated billing features by setting the FeatureSet parameter to CONSOLIDATED_BILLING" , then no policy types are enabled by default and you cannot use organization policies.

Creating a Request

createOrganization :: CreateOrganization

Creates a value of CreateOrganization with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

coFeatureSet - Specifies the feature set supported by the new organization. Each feature set supports different levels of functionality. * CONSOLIDATED_BILLING : All member accounts have their bills consolidated to and paid by the master account. For more information, see Consolidated Billing in the AWS Organizations User Guide . * ALL : In addition to all the features supported by the consolidated billing feature set, the master account can also apply any type of policy to any member account in the organization. For more information, see All features in the AWS Organizations User Guide .

data CreateOrganization

See: createOrganization smart constructor.

Request Lenses

coFeatureSet :: Lens' CreateOrganization (Maybe OrganizationFeatureSet)

Specifies the feature set supported by the new organization. Each feature set supports different levels of functionality. * CONSOLIDATED_BILLING : All member accounts have their bills consolidated to and paid by the master account. For more information, see Consolidated Billing in the AWS Organizations User Guide . * ALL : In addition to all the features supported by the consolidated billing feature set, the master account can also apply any type of policy to any member account in the organization. For more information, see All features in the AWS Organizations User Guide .

Destructuring the Response

createOrganizationResponse

Arguments

:: Int
corsResponseStatus

-> CreateOrganizationResponse

Creates a value of CreateOrganizationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

corsOrganization - A structure that contains details about the newly created organization.

corsResponseStatus - -- | The response status code.

data CreateOrganizationResponse

See: createOrganizationResponse smart constructor.

Response Lenses

corsOrganization :: Lens' CreateOrganizationResponse (Maybe Organization)

A structure that contains details about the newly created organization.

corsResponseStatus :: Lens' CreateOrganizationResponse Int

- | The response status code.