Network.AWS.Organizations.CreateOrganizationalUnit

Description

Creates an organizational unit (OU) within a root or parent OU. An OU is a container for accounts that enables you to organize your accounts to apply policies according to your business requirements. The number of levels deep that you can nest OUs is dependent upon the policy types enabled for that root. For service control policies, the limit is five.

For more information about OUs, see Managing Organizational Units in the AWS Organizations User Guide .

This operation can be called only from the organization's master account.

Creating a Request

createOrganizationalUnit

Arguments

:: Text
couParentId

-> Text
couName

-> CreateOrganizationalUnit

Creates a value of CreateOrganizationalUnit with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

couParentId - The unique identifier (ID) of the parent root or OU in which you want to create the new OU. The regex pattern for a parent ID string requires one of the following: * Root: a string that begins with "r-" followed by from 4 to 32 lower-case letters or digits. * Organizational unit (OU): a string that begins with "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that the OU is in) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.

couName - The friendly name to assign to the new OU.

data CreateOrganizationalUnit

See: createOrganizationalUnit smart constructor.

Request Lenses

couParentId :: Lens' CreateOrganizationalUnit Text

The unique identifier (ID) of the parent root or OU in which you want to create the new OU. The regex pattern for a parent ID string requires one of the following: * Root: a string that begins with "r-" followed by from 4 to 32 lower-case letters or digits. * Organizational unit (OU): a string that begins with "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that the OU is in) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.

couName :: Lens' CreateOrganizationalUnit Text

The friendly name to assign to the new OU.

Destructuring the Response

createOrganizationalUnitResponse

Arguments

:: Int
coursResponseStatus

-> CreateOrganizationalUnitResponse

Creates a value of CreateOrganizationalUnitResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

coursOrganizationalUnit - A structure that contains details about the newly created OU.

coursResponseStatus - -- | The response status code.

data CreateOrganizationalUnitResponse

See: createOrganizationalUnitResponse smart constructor.

Response Lenses

coursOrganizationalUnit :: Lens' CreateOrganizationalUnitResponse (Maybe OrganizationalUnit)

A structure that contains details about the newly created OU.

coursResponseStatus :: Lens' CreateOrganizationalUnitResponse Int

- | The response status code.