Network.AWS.Organizations.DeleteOrganization

Description

Deletes the organization. You can delete an organization only by using credentials from the master account. The organization must be empty of member accounts, OUs, and policies.

Important: If you create any accounts using Organizations operations or the Organizations console, you can't remove those accounts from the organization, which means that you can't delete the organization.

Creating a Request

deleteOrganization :: DeleteOrganization

Creates a value of DeleteOrganization with the minimum fields required to make a request.

data DeleteOrganization

See: deleteOrganization smart constructor.

Destructuring the Response

deleteOrganizationResponse :: DeleteOrganizationResponse

Creates a value of DeleteOrganizationResponse with the minimum fields required to make a request.

data DeleteOrganizationResponse

See: deleteOrganizationResponse smart constructor.