Network.AWS.Organizations.DescribeAccount

Description

Retrieves Organizations-related information about the specified account.

This operation can be called only from the organization's master account.

Creating a Request

describeAccount

Arguments

:: Text
daAccountId

-> DescribeAccount

Creates a value of DescribeAccount with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daAccountId - The unique identifier (ID) of the AWS account that you want information about. You can get the ID from the ListAccounts or ListAccountsForParent operations. The regex pattern for an account ID string requires exactly 12 digits.

data DescribeAccount

See: describeAccount smart constructor.

Request Lenses

daAccountId :: Lens' DescribeAccount Text

The unique identifier (ID) of the AWS account that you want information about. You can get the ID from the ListAccounts or ListAccountsForParent operations. The regex pattern for an account ID string requires exactly 12 digits.

Destructuring the Response

describeAccountResponse

Arguments

:: Int
darsResponseStatus

-> DescribeAccountResponse

Creates a value of DescribeAccountResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

darsAccount - A structure that contains information about the requested account.

darsResponseStatus - -- | The response status code.

data DescribeAccountResponse

See: describeAccountResponse smart constructor.

Response Lenses

darsAccount :: Lens' DescribeAccountResponse (Maybe Account)

A structure that contains information about the requested account.

darsResponseStatus :: Lens' DescribeAccountResponse Int

- | The response status code.