Network.AWS.Organizations.DescribeHandshake

Description

Retrieves information about a previously requested handshake. The handshake ID comes from the response to the original InviteAccountToOrganization operation that generated the handshake.

You can access handshakes that are ACCEPTED, DECLINED, or CANCELED for only 30 days after they change to that state. They are then deleted and no longer accessible.

This operation can be called from any account in the organization.

Creating a Request

describeHandshake

Arguments

:: Text
dhHandshakeId

-> DescribeHandshake

Creates a value of DescribeHandshake with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhHandshakeId - The unique identifier (ID) of the handshake that you want information about. You can get the ID from the original call to InviteAccountToOrganization , or from a call to ListHandshakesForAccount or ListHandshakesForOrganization . The regex pattern for handshake ID string requires "h-" followed by from 8 to 32 lower-case letters or digits.

data DescribeHandshake

See: describeHandshake smart constructor.

Request Lenses

dhHandshakeId :: Lens' DescribeHandshake Text

The unique identifier (ID) of the handshake that you want information about. You can get the ID from the original call to InviteAccountToOrganization , or from a call to ListHandshakesForAccount or ListHandshakesForOrganization . The regex pattern for handshake ID string requires "h-" followed by from 8 to 32 lower-case letters or digits.

Destructuring the Response

describeHandshakeResponse

Arguments

:: Int
dhrsResponseStatus

-> DescribeHandshakeResponse

Creates a value of DescribeHandshakeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhrsHandshake - A structure that contains information about the specified handshake.

dhrsResponseStatus - -- | The response status code.

data DescribeHandshakeResponse

See: describeHandshakeResponse smart constructor.

Response Lenses

dhrsHandshake :: Lens' DescribeHandshakeResponse (Maybe Handshake)

A structure that contains information about the specified handshake.

dhrsResponseStatus :: Lens' DescribeHandshakeResponse Int

- | The response status code.