Network.AWS.Organizations.DescribePolicy

Description

Retrieves information about a policy.

This operation can be called only from the organization's master account.

Creating a Request

describePolicy

Arguments

:: Text
dpPolicyId

-> DescribePolicy

Creates a value of DescribePolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpPolicyId - The unique identifier (ID) of the policy that you want details about. You can get the ID from the ListPolicies or ListPoliciesForTarget operations. The regex pattern for a policy ID string requires "p-" followed by from 8 to 128 lower-case letters or digits.

data DescribePolicy

See: describePolicy smart constructor.

Request Lenses

dpPolicyId :: Lens' DescribePolicy Text

The unique identifier (ID) of the policy that you want details about. You can get the ID from the ListPolicies or ListPoliciesForTarget operations. The regex pattern for a policy ID string requires "p-" followed by from 8 to 128 lower-case letters or digits.

Destructuring the Response

describePolicyResponse

Arguments

:: Int
dprsResponseStatus

-> DescribePolicyResponse

Creates a value of DescribePolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dprsPolicy - A structure that contains details about the specified policy.

dprsResponseStatus - -- | The response status code.

data DescribePolicyResponse

See: describePolicyResponse smart constructor.

Response Lenses

dprsPolicy :: Lens' DescribePolicyResponse (Maybe Policy)

A structure that contains details about the specified policy.

dprsResponseStatus :: Lens' DescribePolicyResponse Int

- | The response status code.