Network.AWS.Organizations.DetachPolicy

Description

Detaches a policy from a target root, organizational unit, or account. If the policy being detached is a service control policy (SCP), the changes to permissions for IAM users and roles in affected accounts are immediate.

Note: Every root, OU, and account must have at least one SCP attached. If you want to replace the default FullAWSAccess policy with one that limits the permissions that can be delegated, then you must attach the replacement policy before you can remove the default one. This is the authorization strategy of whitelisting . If you instead attach a second SCP and leave the FullAWSAccess SCP still attached, and specify Effect: Deny in the second SCP to override the Effect: Allow in the FullAWSAccess policy (or any other attached SCP), then you are using the authorization strategy of blacklisting .

This operation can be called only from the organization's master account.

Creating a Request

detachPolicy

Arguments

:: Text
detPolicyId

-> Text
detTargetId

-> DetachPolicy

Creates a value of DetachPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

detPolicyId - The unique identifier (ID) of the policy you want to detach. You can get the ID from the ListPolicies or ListPoliciesForTarget operations. The regex pattern for a policy ID string requires "p-" followed by from 8 to 128 lower-case letters or digits.

detTargetId - The unique identifier (ID) of the root, OU, or account from which you want to detach the policy. You can get the ID from the ListRoots , ListOrganizationalUnitsForParent , or ListAccounts operations. The regex pattern for a target ID string requires one of the following: * Root: a string that begins with "r-" followed by from 4 to 32 lower-case letters or digits. * Account: a string that consists of exactly 12 digits. * Organizational unit (OU): a string that begins with "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that the OU is in) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.

data DetachPolicy

See: detachPolicy smart constructor.

Request Lenses

detPolicyId :: Lens' DetachPolicy Text

The unique identifier (ID) of the policy you want to detach. You can get the ID from the ListPolicies or ListPoliciesForTarget operations. The regex pattern for a policy ID string requires "p-" followed by from 8 to 128 lower-case letters or digits.

detTargetId :: Lens' DetachPolicy Text

The unique identifier (ID) of the root, OU, or account from which you want to detach the policy. You can get the ID from the ListRoots , ListOrganizationalUnitsForParent , or ListAccounts operations. The regex pattern for a target ID string requires one of the following: * Root: a string that begins with "r-" followed by from 4 to 32 lower-case letters or digits. * Account: a string that consists of exactly 12 digits. * Organizational unit (OU): a string that begins with "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that the OU is in) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.

Destructuring the Response

detachPolicyResponse :: DetachPolicyResponse

Creates a value of DetachPolicyResponse with the minimum fields required to make a request.

data DetachPolicyResponse

See: detachPolicyResponse smart constructor.