Network.AWS.Organizations.DisablePolicyType

Description

Disables an organizational control policy type in a root. A poicy of a certain type can be attached to entities in a root only if that type is enabled in the root. After you perform this operation, you no longer can attach policies of the specified type to that root or to any OU or account in that root. You can undo this by using the EnablePolicyType operation.

This operation can be called only from the organization's master account.

Creating a Request

disablePolicyType

Arguments

:: Text
dptRootId

-> PolicyType
dptPolicyType

-> DisablePolicyType

Creates a value of DisablePolicyType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dptRootId - The unique identifier (ID) of the root in which you want to disable a policy type. You can get the ID from the ListPolicies operation. The regex pattern for a root ID string requires "r-" followed by from 4 to 32 lower-case letters or digits.

dptPolicyType - The policy type that you want to disable in this root.

data DisablePolicyType

See: disablePolicyType smart constructor.

Request Lenses

dptRootId :: Lens' DisablePolicyType Text

The unique identifier (ID) of the root in which you want to disable a policy type. You can get the ID from the ListPolicies operation. The regex pattern for a root ID string requires "r-" followed by from 4 to 32 lower-case letters or digits.

dptPolicyType :: Lens' DisablePolicyType PolicyType

The policy type that you want to disable in this root.

Destructuring the Response

disablePolicyTypeResponse

Arguments

:: Int
dptrsResponseStatus

-> DisablePolicyTypeResponse

Creates a value of DisablePolicyTypeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dptrsRoot - A structure that shows the root with the updated list of enabled policy types.

dptrsResponseStatus - -- | The response status code.

data DisablePolicyTypeResponse

See: disablePolicyTypeResponse smart constructor.

Response Lenses

dptrsRoot :: Lens' DisablePolicyTypeResponse (Maybe Root)

A structure that shows the root with the updated list of enabled policy types.

dptrsResponseStatus :: Lens' DisablePolicyTypeResponse Int

- | The response status code.