Network.AWS.Organizations.EnablePolicyType

Description

Enables a policy type in a root. After you enable a policy type in a root, you can attach policies of that type to the root, any OU, or account in that root. You can undo this by using the DisablePolicyType operation.

This operation can be called only from the organization's master account.

Creating a Request

enablePolicyType

Arguments

:: Text
eptRootId

-> PolicyType
eptPolicyType

-> EnablePolicyType

Creates a value of EnablePolicyType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eptRootId - The unique identifier (ID) of the root in which you want to enable a policy type. You can get the ID from the ListRoots operation. The regex pattern for a root ID string requires "r-" followed by from 4 to 32 lower-case letters or digits.

eptPolicyType - The policy type that you want to enable.

data EnablePolicyType

See: enablePolicyType smart constructor.

Request Lenses

eptRootId :: Lens' EnablePolicyType Text

The unique identifier (ID) of the root in which you want to enable a policy type. You can get the ID from the ListRoots operation. The regex pattern for a root ID string requires "r-" followed by from 4 to 32 lower-case letters or digits.

eptPolicyType :: Lens' EnablePolicyType PolicyType

The policy type that you want to enable.

Destructuring the Response

enablePolicyTypeResponse

Arguments

:: Int
eptrsResponseStatus

-> EnablePolicyTypeResponse

Creates a value of EnablePolicyTypeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eptrsRoot - A structure that shows the root with the updated list of enabled policy types.

eptrsResponseStatus - -- | The response status code.

data EnablePolicyTypeResponse

See: enablePolicyTypeResponse smart constructor.

Response Lenses

eptrsRoot :: Lens' EnablePolicyTypeResponse (Maybe Root)

A structure that shows the root with the updated list of enabled policy types.

eptrsResponseStatus :: Lens' EnablePolicyTypeResponse Int

- | The response status code.