Network.AWS.Organizations.LeaveOrganization

Description

Removes a member account from its parent organization. This version of the operation is performed by the account that wants to leave. To remove a member account as a user in the master account, use RemoveAccountFromOrganization instead.

This operation can be called only from a member account in the organization.

Important: * The master account in an organization with all features enabled can set service control policies (SCPs) that can restrict what administrators of member accounts can do, including preventing them from successfully calling LeaveOrganization and leaving the organization.

If you created the account using the AWS Organizations console, the Organizations API, or the Organizations CLI commands, then you cannot remove the account.

You can leave an organization only after you enable IAM user access to billing in your account. For more information, see Activating Access to the Billing and Cost Management Console in the AWS Billing and Cost Management User Guide .

Creating a Request

leaveOrganization :: LeaveOrganization

Creates a value of LeaveOrganization with the minimum fields required to make a request.

data LeaveOrganization

See: leaveOrganization smart constructor.

Destructuring the Response

leaveOrganizationResponse :: LeaveOrganizationResponse

Creates a value of LeaveOrganizationResponse with the minimum fields required to make a request.

data LeaveOrganizationResponse

See: leaveOrganizationResponse smart constructor.