Network.AWS.Organizations.ListAccountsForParent

Description

Lists the accounts in an organization that are contained by the specified target root or organizational unit (OU). If you specify the root, you get a list of all the accounts that are not in any OU. If you specify an OU, you get a list of all the accounts in only that OU, and not in any child OUs. To get a list of all accounts in the organization, use the ListAccounts operation.

This operation returns paginated results.

Creating a Request

listAccountsForParent

Arguments

:: Text
lafpParentId

-> ListAccountsForParent

Creates a value of ListAccountsForParent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lafpNextToken - Use this parameter if you receive a NextToken response in a previous request that indicates that there is more output available. Set it to the value of the previous call's NextToken response to indicate where the output should continue from.

lafpMaxResults - (Optional) Use this to limit the number of results you want included in the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

lafpParentId - The unique identifier (ID) for the parent root or organization unit (OU) whose accounts you want to list.

data ListAccountsForParent

See: listAccountsForParent smart constructor.

Request Lenses

lafpNextToken :: Lens' ListAccountsForParent (Maybe Text)

Use this parameter if you receive a NextToken response in a previous request that indicates that there is more output available. Set it to the value of the previous call's NextToken response to indicate where the output should continue from.

lafpMaxResults :: Lens' ListAccountsForParent (Maybe Natural)

(Optional) Use this to limit the number of results you want included in the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

lafpParentId :: Lens' ListAccountsForParent Text

The unique identifier (ID) for the parent root or organization unit (OU) whose accounts you want to list.

Destructuring the Response

listAccountsForParentResponse

Arguments

:: Int
lafprsResponseStatus

-> ListAccountsForParentResponse

Creates a value of ListAccountsForParentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lafprsAccounts - A list of the accounts in the specified root or OU.

lafprsNextToken - If present, this value indicates that there is more output available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null .

lafprsResponseStatus - -- | The response status code.

data ListAccountsForParentResponse

See: listAccountsForParentResponse smart constructor.

Response Lenses

lafprsAccounts :: Lens' ListAccountsForParentResponse [Account]

A list of the accounts in the specified root or OU.

lafprsNextToken :: Lens' ListAccountsForParentResponse (Maybe Text)

If present, this value indicates that there is more output available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null .

lafprsResponseStatus :: Lens' ListAccountsForParentResponse Int

- | The response status code.