Network.AWS.Organizations.ListChildren

Description

Lists all of the OUs or accounts that are contained in the specified parent OU or root. This operation, along with ListParents enables you to traverse the tree structure that makes up this root.

This operation returns paginated results.

Creating a Request

listChildren

Arguments

:: Text
lcParentId

-> ChildType
lcChildType

-> ListChildren

Creates a value of ListChildren with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcNextToken - Use this parameter if you receive a NextToken response in a previous request that indicates that there is more output available. Set it to the value of the previous call's NextToken response to indicate where the output should continue from.

lcMaxResults - (Optional) Use this to limit the number of results you want included in the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

lcParentId - The unique identifier (ID) for the parent root or OU whose children you want to list. The regex pattern for a parent ID string requires one of the following: * Root: a string that begins with "r-" followed by from 4 to 32 lower-case letters or digits. * Organizational unit (OU): a string that begins with "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that the OU is in) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.

lcChildType - Filters the output to include only the specified child type.

data ListChildren

See: listChildren smart constructor.

Request Lenses

lcNextToken :: Lens' ListChildren (Maybe Text)

Use this parameter if you receive a NextToken response in a previous request that indicates that there is more output available. Set it to the value of the previous call's NextToken response to indicate where the output should continue from.

lcMaxResults :: Lens' ListChildren (Maybe Natural)

(Optional) Use this to limit the number of results you want included in the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

lcParentId :: Lens' ListChildren Text

The unique identifier (ID) for the parent root or OU whose children you want to list. The regex pattern for a parent ID string requires one of the following: * Root: a string that begins with "r-" followed by from 4 to 32 lower-case letters or digits. * Organizational unit (OU): a string that begins with "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that the OU is in) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.

lcChildType :: Lens' ListChildren ChildType

Filters the output to include only the specified child type.

Destructuring the Response

listChildrenResponse

Arguments

:: Int
lcrsResponseStatus

-> ListChildrenResponse

Creates a value of ListChildrenResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcrsChildren - The list of children of the specified parent container.

lcrsNextToken - If present, this value indicates that there is more output available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null .

lcrsResponseStatus - -- | The response status code.

data ListChildrenResponse

See: listChildrenResponse smart constructor.

Response Lenses

lcrsChildren :: Lens' ListChildrenResponse [Child]

The list of children of the specified parent container.

lcrsNextToken :: Lens' ListChildrenResponse (Maybe Text)

If present, this value indicates that there is more output available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null .

lcrsResponseStatus :: Lens' ListChildrenResponse Int

- | The response status code.