Network.AWS.Organizations.ListHandshakesForAccount

Description

Lists the current handshakes that are associated with the account of the requesting user.

Handshakes that are ACCEPTED, DECLINED, or CANCELED appear in the results of this API for only 30 days after changing to that state. After that they are deleted and no longer accessible.

This operation can be called from any account in the organization.

This operation returns paginated results.

Creating a Request

listHandshakesForAccount :: ListHandshakesForAccount

Creates a value of ListHandshakesForAccount with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lhfaNextToken - Use this parameter if you receive a NextToken response in a previous request that indicates that there is more output available. Set it to the value of the previous call's NextToken response to indicate where the output should continue from.

lhfaFilter - Filters the handshakes that you want included in the response. The default is all types. Use the ActionType element to limit the output to only a specified type, such as INVITE , ENABLE-FULL-CONTROL , or APPROVE-FULL-CONTROL . Alternatively, for the ENABLE-FULL-CONTROL handshake that generates a separate child handshake for each member account, you can specify ParentHandshakeId to see only the handshakes that were generated by that parent request.

lhfaMaxResults - (Optional) Use this to limit the number of results you want included in the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

data ListHandshakesForAccount

See: listHandshakesForAccount smart constructor.

Request Lenses

lhfaNextToken :: Lens' ListHandshakesForAccount (Maybe Text)

Use this parameter if you receive a NextToken response in a previous request that indicates that there is more output available. Set it to the value of the previous call's NextToken response to indicate where the output should continue from.

lhfaFilter :: Lens' ListHandshakesForAccount (Maybe HandshakeFilter)

Filters the handshakes that you want included in the response. The default is all types. Use the ActionType element to limit the output to only a specified type, such as INVITE , ENABLE-FULL-CONTROL , or APPROVE-FULL-CONTROL . Alternatively, for the ENABLE-FULL-CONTROL handshake that generates a separate child handshake for each member account, you can specify ParentHandshakeId to see only the handshakes that were generated by that parent request.

lhfaMaxResults :: Lens' ListHandshakesForAccount (Maybe Natural)

(Optional) Use this to limit the number of results you want included in the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

Destructuring the Response

listHandshakesForAccountResponse

Arguments

:: Int
lhfarsResponseStatus

-> ListHandshakesForAccountResponse

Creates a value of ListHandshakesForAccountResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lhfarsHandshakes - A list of Handshake objects with details about each of the handshakes that is associated with the specified account.

lhfarsNextToken - If present, this value indicates that there is more output available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null .

lhfarsResponseStatus - -- | The response status code.

data ListHandshakesForAccountResponse

See: listHandshakesForAccountResponse smart constructor.

Response Lenses

lhfarsHandshakes :: Lens' ListHandshakesForAccountResponse [Handshake]

A list of Handshake objects with details about each of the handshakes that is associated with the specified account.

lhfarsNextToken :: Lens' ListHandshakesForAccountResponse (Maybe Text)

If present, this value indicates that there is more output available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null .

lhfarsResponseStatus :: Lens' ListHandshakesForAccountResponse Int

- | The response status code.