Network.AWS.Organizations.ListOrganizationalUnitsForParent

Description

Lists the organizational units (OUs) in a parent organizational unit or root.

This operation can be called only from the organization's master account.

This operation returns paginated results.

Creating a Request

listOrganizationalUnitsForParent

Arguments

:: Text
loufpParentId

-> ListOrganizationalUnitsForParent

Creates a value of ListOrganizationalUnitsForParent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

loufpNextToken - Use this parameter if you receive a NextToken response in a previous request that indicates that there is more output available. Set it to the value of the previous call's NextToken response to indicate where the output should continue from.

loufpMaxResults - (Optional) Use this to limit the number of results you want included in the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

loufpParentId - The unique identifier (ID) of the root or OU whose child OUs you want to list. The regex pattern for a parent ID string requires one of the following: * Root: a string that begins with "r-" followed by from 4 to 32 lower-case letters or digits. * Organizational unit (OU): a string that begins with "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that the OU is in) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.

data ListOrganizationalUnitsForParent

See: listOrganizationalUnitsForParent smart constructor.

Request Lenses

loufpNextToken :: Lens' ListOrganizationalUnitsForParent (Maybe Text)

Use this parameter if you receive a NextToken response in a previous request that indicates that there is more output available. Set it to the value of the previous call's NextToken response to indicate where the output should continue from.

loufpMaxResults :: Lens' ListOrganizationalUnitsForParent (Maybe Natural)

(Optional) Use this to limit the number of results you want included in the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

loufpParentId :: Lens' ListOrganizationalUnitsForParent Text

The unique identifier (ID) of the root or OU whose child OUs you want to list. The regex pattern for a parent ID string requires one of the following: * Root: a string that begins with "r-" followed by from 4 to 32 lower-case letters or digits. * Organizational unit (OU): a string that begins with "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that the OU is in) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.

Destructuring the Response

listOrganizationalUnitsForParentResponse

Arguments

:: Int
loufprsResponseStatus

-> ListOrganizationalUnitsForParentResponse

Creates a value of ListOrganizationalUnitsForParentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

loufprsNextToken - If present, this value indicates that there is more output available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null .

loufprsOrganizationalUnits - A list of the OUs in the specified root or parent OU.

loufprsResponseStatus - -- | The response status code.

data ListOrganizationalUnitsForParentResponse

See: listOrganizationalUnitsForParentResponse smart constructor.

Response Lenses

loufprsNextToken :: Lens' ListOrganizationalUnitsForParentResponse (Maybe Text)

If present, this value indicates that there is more output available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null .

loufprsOrganizationalUnits :: Lens' ListOrganizationalUnitsForParentResponse [OrganizationalUnit]

A list of the OUs in the specified root or parent OU.

loufprsResponseStatus :: Lens' ListOrganizationalUnitsForParentResponse Int

- | The response status code.