Network.AWS.Organizations.ListTargetsForPolicy

Description

Lists all the roots, OUs, and accounts to which the specified policy is attached.

This operation can be called only from the organization's master account.

This operation returns paginated results.

Creating a Request

listTargetsForPolicy

Arguments

:: Text
ltfpPolicyId

-> ListTargetsForPolicy

Creates a value of ListTargetsForPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltfpNextToken - Use this parameter if you receive a NextToken response in a previous request that indicates that there is more output available. Set it to the value of the previous call's NextToken response to indicate where the output should continue from.

ltfpMaxResults - (Optional) Use this to limit the number of results you want included in the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

ltfpPolicyId - The unique identifier (ID) of the policy for which you want to know its attachments. The regex pattern for a policy ID string requires "p-" followed by from 8 to 128 lower-case letters or digits.

data ListTargetsForPolicy

See: listTargetsForPolicy smart constructor.

Request Lenses

ltfpNextToken :: Lens' ListTargetsForPolicy (Maybe Text)

Use this parameter if you receive a NextToken response in a previous request that indicates that there is more output available. Set it to the value of the previous call's NextToken response to indicate where the output should continue from.

ltfpMaxResults :: Lens' ListTargetsForPolicy (Maybe Natural)

(Optional) Use this to limit the number of results you want included in the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

ltfpPolicyId :: Lens' ListTargetsForPolicy Text

The unique identifier (ID) of the policy for which you want to know its attachments. The regex pattern for a policy ID string requires "p-" followed by from 8 to 128 lower-case letters or digits.

Destructuring the Response

listTargetsForPolicyResponse

Arguments

:: Int
ltfprsResponseStatus

-> ListTargetsForPolicyResponse

Creates a value of ListTargetsForPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltfprsNextToken - If present, this value indicates that there is more output available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null .

ltfprsTargets - A list of structures, each of which contains details about one of the entities to which the specified policy is attached.

ltfprsResponseStatus - -- | The response status code.

data ListTargetsForPolicyResponse

See: listTargetsForPolicyResponse smart constructor.

Response Lenses

ltfprsNextToken :: Lens' ListTargetsForPolicyResponse (Maybe Text)

If present, this value indicates that there is more output available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null .

ltfprsTargets :: Lens' ListTargetsForPolicyResponse [PolicyTargetSummary]

A list of structures, each of which contains details about one of the entities to which the specified policy is attached.

ltfprsResponseStatus :: Lens' ListTargetsForPolicyResponse Int

- | The response status code.