Network.AWS.Organizations.MoveAccount

Description

Moves an account from its current source parent root or OU to the specified destination parent root or OU.

This operation can be called only from the organization's master account.

Creating a Request

moveAccount

Arguments

:: Text
maAccountId

-> Text
maSourceParentId

-> Text
maDestinationParentId

-> MoveAccount

Creates a value of MoveAccount with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

maAccountId - The unique identifier (ID) of the account that you want to move. The regex pattern for an account ID string requires exactly 12 digits.

maSourceParentId - The unique identifier (ID) of the root or organizational unit that you want to move the account from. The regex pattern for a parent ID string requires one of the following: * Root: a string that begins with "r-" followed by from 4 to 32 lower-case letters or digits. * Organizational unit (OU): a string that begins with "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that the OU is in) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.

maDestinationParentId - The unique identifier (ID) of the root or organizational unit that you want to move the account to. The regex pattern for a parent ID string requires one of the following: * Root: a string that begins with "r-" followed by from 4 to 32 lower-case letters or digits. * Organizational unit (OU): a string that begins with "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that the OU is in) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.

data MoveAccount

See: moveAccount smart constructor.

Request Lenses

maAccountId :: Lens' MoveAccount Text

The unique identifier (ID) of the account that you want to move. The regex pattern for an account ID string requires exactly 12 digits.

maSourceParentId :: Lens' MoveAccount Text

The unique identifier (ID) of the root or organizational unit that you want to move the account from. The regex pattern for a parent ID string requires one of the following: * Root: a string that begins with "r-" followed by from 4 to 32 lower-case letters or digits. * Organizational unit (OU): a string that begins with "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that the OU is in) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.

maDestinationParentId :: Lens' MoveAccount Text

The unique identifier (ID) of the root or organizational unit that you want to move the account to. The regex pattern for a parent ID string requires one of the following: * Root: a string that begins with "r-" followed by from 4 to 32 lower-case letters or digits. * Organizational unit (OU): a string that begins with "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that the OU is in) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.

Destructuring the Response

moveAccountResponse :: MoveAccountResponse

Creates a value of MoveAccountResponse with the minimum fields required to make a request.

data MoveAccountResponse

See: moveAccountResponse smart constructor.