Network.AWS.Organizations.UpdateOrganizationalUnit

Description

Renames the specified organizational unit (OU). The ID and ARN do not change. The child OUs and accounts remain in place, and any attached policies of the OU remain attached.

This operation can be called only from the organization's master account.

Creating a Request

updateOrganizationalUnit

Arguments

:: Text
uouOrganizationalUnitId

-> UpdateOrganizationalUnit

Creates a value of UpdateOrganizationalUnit with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uouName - The new name that you want to assign to the OU. The regex pattern that is used to validate this parameter is a string of any of the characters in the ASCII character range.

uouOrganizationalUnitId - The unique identifier (ID) of the OU that you want to rename. You can get the ID from the ListOrganizationalUnitsForParent operation. The regex pattern for an organizational unit ID string requires "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that contains the OU) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.

data UpdateOrganizationalUnit

See: updateOrganizationalUnit smart constructor.

Request Lenses

uouName :: Lens' UpdateOrganizationalUnit (Maybe Text)

The new name that you want to assign to the OU. The regex pattern that is used to validate this parameter is a string of any of the characters in the ASCII character range.

uouOrganizationalUnitId :: Lens' UpdateOrganizationalUnit Text

The unique identifier (ID) of the OU that you want to rename. You can get the ID from the ListOrganizationalUnitsForParent operation. The regex pattern for an organizational unit ID string requires "ou-" followed by from 4 to 32 lower-case letters or digits (the ID of the root that contains the OU) followed by a second "-" dash and from 8 to 32 additional lower-case letters or digits.

Destructuring the Response

updateOrganizationalUnitResponse

Arguments

:: Int
uoursResponseStatus

-> UpdateOrganizationalUnitResponse

Creates a value of UpdateOrganizationalUnitResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uoursOrganizationalUnit - A structure that contains the details about the specified OU, including its new name.

uoursResponseStatus - -- | The response status code.

data UpdateOrganizationalUnitResponse

See: updateOrganizationalUnitResponse smart constructor.

Response Lenses

uoursOrganizationalUnit :: Lens' UpdateOrganizationalUnitResponse (Maybe OrganizationalUnit)

A structure that contains the details about the specified OU, including its new name.

uoursResponseStatus :: Lens' UpdateOrganizationalUnitResponse Int

- | The response status code.